--- a/NBitcoin/Base58Data.cs
+++ b/NBitcoin/Base58Data.cs
@@ -22,8 +22,13 @@ namespace NBitcoin
 	/// </summary>
 	public abstract class Base58Data : IBase58Data
 	{
+#if HAS_SPAN
+		protected byte[] vchData = Array.Empty<byte>();
+		protected ReadOnlyMemory<byte> vchVersion;
+#else
 		protected byte[] vchData = new byte[0];
 		protected byte[] vchVersion = new byte[0];
+#endif
 		protected string wifData = "";
 		private Network _Network;
 		public Network Network
@@ -61,11 +66,26 @@ namespace NBitcoin
 			}
 
 			byte[] vchTemp = _Network.NetworkStringParser.GetBase58CheckEncoder().DecodeData(psz);
-			var expectedVersion = _Network.GetVersionBytes(Type, true);
+#if HAS_SPAN
+			if (!(_Network.GetVersionMemory(Type, false) is ReadOnlyMemory<byte> expectedVersion))
+				throw new FormatException("Invalid " + this.GetType().Name);
+#else
+			var expectedVersion = _Network.GetVersionBytes(Type, false);
+			if (expectedVersion is null)
+				throw new FormatException("Invalid " + this.GetType().Name);
+#endif
 
-
+#if HAS_SPAN
+			var vchTempMemory = vchTemp.AsMemory();
+			vchVersion = vchTempMemory.Slice(0, expectedVersion.Length);
+#else
 			vchVersion = vchTemp.SafeSubarray(0, expectedVersion.Length);
+#endif
+#if HAS_SPAN
+			if (!vchVersion.Span.SequenceEqual(expectedVersion.Span))
+#else
 			if (!Utils.ArrayEqual(vchVersion, expectedVersion))
+#endif
 			{
 				if (_Network.NetworkStringParser.TryParse(psz, Network, out T other))
 				{
@@ -80,7 +100,11 @@ namespace NBitcoin
 			}
 			else
 			{
+#if HAS_SPAN
+				vchData = vchTempMemory.Slice(expectedVersion.Length).ToArray();
+#else
 				vchData = vchTemp.SafeSubarray(expectedVersion.Length);
+#endif
 				wifData = psz;
 			}
 
@@ -93,8 +117,19 @@ namespace NBitcoin
 		private void SetData(byte[] vchData)
 		{
 			this.vchData = vchData;
-			this.vchVersion = _Network.GetVersionBytes(Type, true);
+#if HAS_SPAN
+			if (!(_Network.GetVersionMemory(Type, false) is ReadOnlyMemory<byte> v))
+				throw new FormatException("Invalid " + this.GetType().Name);
+			this.vchVersion = v;
+			Span<byte> buffer = vchVersion.Length + vchData.Length is int length &&
+								length > 256 ? new byte[length] : stackalloc byte[length];
+			this.vchVersion.Span.CopyTo(buffer);
+			this.vchData.CopyTo(buffer.Slice(this.vchVersion.Length));
+			wifData = _Network.NetworkStringParser.GetBase58CheckEncoder().EncodeData(buffer);
+#else
+			this.vchVersion = _Network.GetVersionBytes(Type, false);
 			wifData = _Network.NetworkStringParser.GetBase58CheckEncoder().EncodeData(vchVersion.Concat(vchData).ToArray());
+#endif
 
 			if (!IsValid)
 				throw new FormatException("Invalid " + this.GetType().Name);