It turns out that we were de serializing twice in the address parsing logic.

Before

|            Method |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |----------:|----------:|----------:|-------:|------:|------:|----------:|
| DeserializeBase58 |  9.638 us | 0.1620 us | 0.1353 us | 0.2747 |     - |     - |   2.26 KB |
|   DeserializeBech |  3.593 us | 0.0146 us | 0.0137 us | 0.1335 |     - |     - |   1.11 KB |
|   DeserializeXPub | 30.870 us | 0.3133 us | 0.2616 us | 0.1831 |     - |     - |   1.97 KB

After
|            Method |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |----------:|----------:|----------:|-------:|------:|------:|----------:|
| DeserializeBase58 |  5.308 us | 0.0504 us | 0.0393 us | 0.1526 |     - |     - |    1328 B |
|   DeserializeBech |  1.445 us | 0.0100 us | 0.0089 us | 0.0420 |     - |     - |     360 B |
|   DeserializeXPub | 32.865 us | 0.3623 us | 0.3389 us | 0.1831 |     - |     - |    1912 B |