@NicolasDorier 2020-05-25
I think I will stop perf improvement on string parsing for now... this is pretty useless to optimize more. It is 1 000 000 operations per second for bech32.